bring over stuff from phoo
but something got broken. the "online" list isn't being updated right.

+++ selfs/MsgView.qml
@@ -25,6 +25,7 @@ Page {
     property bool hid
     property string text_bg_color: primary_dark
     property string export_result
+    property bool is_trash: false
 
     //anchors.fill:parent
 
@@ -43,6 +44,7 @@ Page {
             core.stop_zap_view(view_id)
             core.delete_zap_view(view_id)
             }
+            is_trash = false
         }
         else
         {
@@ -79,6 +81,18 @@ Page {
             fav = core.has_tag_message(mid, "_fav")
             hid = core.has_tag_message(mid, "_hid")
         }
+        function onQt_app_state_change(app_state) {
+            if(app_state === 0) {
+
+
+            }
+            if(app_state !== 0) {
+                if(view_id !== -1) {
+                core.stop_zap_view(view_id)
+                core.delete_zap_view(view_id)
+                }
+            }
+        }
     }
 
     Component {
@@ -90,21 +104,50 @@ Page {
                 source: mi("ic_action_overflow.png")
             }
             onClicked: optionsMenu.open()
+            MenuItem {
+                id: untrash_item
+                text: "Untrash msg"
+                visible: false
+                onTriggered: {
+                    core.unset_tag_message(mid, "_trash")
+                    themsglist.invalidate_model_filter()
+                    stack.pop()
+                }
+            }
+            MenuItem {
+                id: review_item
+                text: "Review"
+                visible: false
+                onTriggered: {
+                    stack.push(msg_review)
 
+                }
+            }
             Menu {
                 id: optionsMenu
                 x: parent.width - width
                 transformOrigin: Menu.TopRight
                 MenuItem {
-                    text: "Delete msg"
+                    text: is_trash ? "Delete forever" : "Trash msg"
                     onTriggered: {
-                        core.delete_message(uid, mid)
-                        themsglist.reload_model()
+                        if(is_trash) { 
+                            core.delete_message(uid, mid)
+                        } else {
+                            core.set_tag_message(mid, "_trash")
+                        }
+                        themsglist.invalidate_model_filter()
                         stack.pop()
-
                     }
                 }
-
+//                MenuItem {
+//                    text: "Untrash msg"
+//                    visible: is_trash
+//                    onTriggered: {
+//                        core.unset_tag_message(mid, "_trash")
+//                        themsglist.invalidate_model_filter()
+//                        stack.pop()
+//                    }
+//                }
                 MenuItem {
                     text: "Forward msg"
                     onTriggered: {
@@ -127,6 +170,7 @@ Page {
                             core.unset_tag_message(mid, "_hid")
                         else
                             core.set_tag_message(mid, "_hid")
+                        themsglist.invalidate_model_filter()
                     }
                 }
                 MenuItem {
@@ -145,15 +189,35 @@ Page {
                     }
                 }
 
-                MenuItem {
-                    text: "Review"
-                    visible: core.this_uid === the_man
-                    onTriggered: {
-                        stack.push(msg_review)
+//                MenuItem {
+//                    text: "Review"
+//                    visible: core.this_uid === the_man
+//                    onTriggered: {
+//                        stack.push(msg_review)
 
+//                    }
+//                }
+
+
+                onVisibleChanged: {
+                    if(visible) {
+                        if(is_trash) {
+                            untrash_item.visible = true
+                            optionsMenu.insertItem(0, untrash_item)
+                        }
+                        else {
+                            untrash_item.visible = false
+                            optionsMenu.removeItem(untrash_item)
+                        }
+                        if(core.this_uid === the_man) {
+                            review_item.visible = true
+                            optionsMenu.addItem(review_item)
+                        } else {
+                            review_item.visible = false
+                            optionsMenu.removeItem(review_item)
+                        }
                     }
                 }
-
 
             }
         }
